Added MorphologicalProcessor class for morphological operations
