+++ CMakeLists.txt
@@ -18,6 +18,8 @@ add_library(image_processing_lib
         include/InputProcessor.h
         src/InputProcessor.cpp
         src/ImageProcessor.cpp
+        src/MorphologicalProcessor.cpp
+        include/MorphologicalProcessor.h
 )
 
 target_link_libraries(image_processing_lib PUBLIC ${OpenCV_LIBS})
